Paramati
Coding Assignment

******************************

1. Design a Java project/ program which will take a Text file as an input and:

a. Be able to calculate the no of characters, words, and sentences in that file.

b. Be able to search for a given pattern in that file giving out information like the line nos. of all matching entries.

For Part b, a pattern(for simplicity, let us assume that it would be a continuous string of characters) will be input by the user which needs to be matched across the text file that was input to the program/ application.

A hint is if you have used any Text tools like TextPad or NotePad++, they have this pattern search. Ideally, regex also works as part of pattern, but since this is a quick coding assignment, we will skip regex for now.